requirements.txt
torch>=1.13.0
transformers>=4.26.0
opencv-python
Pillow
tqdm